moved most of app js code to new component workflow designer and formatted all code
